+++ slide/flip/amd/src/slide.js
@@ -110,6 +110,10 @@ export default class Slide extends BaseSlide {
         const target = e.target.closest(SELECTORS.listenItem);
         var cardBlock = target.querySelector(SELECTORS.flipCardBlock)
 
+        if (window.NctSlidesFX) {
+            window.NctSlidesFX.play('flip');
+        }
+
         const anim = cardBlock.getAnimations()?.[0];
         if (cardBlock.dataset.flipped == 'true') {
             anim.play();